test(agent): split responder snapshot-cache/budget unit tests into their own file (#1594)
Follow-up to #1569 addressing the post-merge review note that the responder
interleaving test file crossed the 1k-line threshold with unrelated concerns.

Move the snapshot-memo and shared-budget UNIT tests (and the budget-defaults
constant checks) out of sync-responder-concurrent-interleaving.test.ts into a new
sync-responder-snapshot-cache.test.ts, so cache/budget behaviour lives next to its
abstraction and the interleaving file holds only handler-level pagination
scenarios. No assertions change; the full agent unit suite still runs 461 tests.
The interleaving file drops from ~1,493 to ~890 lines.
